Fix up the windows platform path
Some of these changes have had me question my sanity beyond belief.

TokenType was clashing with an enum value in winnt.h, even if I
uninclude literally every single windows header. Must be brought in by
std somewhere. Fix: move the TokenType enum into the Token struct,
rename it Token. Plus side, slightly nicer syntax to refer to the
different token types.

Matrix4f::orthographic had an error that literally said "syntax error: )"
The problem? Windows defines near and far to some macro which was
clashing with the near and far function parameter. Fix: undef

Windows, go fuck yourself

+++ src/leary.cpp
@@ -203,7 +203,7 @@ void game_render(GameState *game)
 	render_info.framebuffer       = game->vulkan.framebuffers[image_index];
 	render_info.renderArea.offset = { 0, 0 };
 	render_info.renderArea.extent = game->vulkan.swapchain.extent;
-	render_info.clearValueCount   = clear_values.size();
+	render_info.clearValueCount   = (u32)clear_values.size();
 	render_info.pClearValues      = clear_values.data();
 
 	VkDeviceSize offsets[] = { 0 };
